refactor: changed the way the global styles are imported and created the wrappers folder inserting the corresponding files in it
