+++ pages/_app.tsx
@@ -3,21 +3,13 @@ import { ThemeProvider } from '@mui/system'
 import { theme } from 'ui/_styles'
 import { CssBaseline } from '@mui/material'
 import { appWithTranslation } from 'next-i18next'
+import '../public/styles/common.css'
 
 function MyApp({ Component, pageProps }: AppProps) {
   return (
     <ThemeProvider theme={theme}>
       <CssBaseline />
       <Component {...pageProps} />
-      <style jsx global>{`
-        html,
-        body {
-          height: 100%;
-        }
-        #__next {
-          height: 100%;
-        }
-      `}</style>
     </ThemeProvider>
   )
 }
