+++ public/styles/common.css
@@ -1,0 +1,7 @@
+html,
+body {
+  height: 100%;
+}
+#__next {
+  height: 100%;
+}
