+++ src/ui/_components/index.ts
@@ -1,4 +1,3 @@
-/* eslint-disable prettier/prettier */
 export { Text } from './atoms/Text'
 
 export { Box } from './objects/Box'
@@ -6,6 +5,6 @@ export { Row } from './objects/Row'
 export { CommonGoldInput } from './objects/inputs/CommonGoldInputs'
 export { BigGoldInput } from './objects/inputs/BigGoldInput'
 export { GoldSelectBox } from './objects/inputs/GoldSelectBox'
-export { FormWrapper } from './objects/inputs/FormWrapper'
-export { ContentWrapper } from './objects/inputs/ContentWrapper'
-export { HeaderWrapper } from './objects/inputs/HeaderWrapper'
+export { FormWrapper } from './objects/wrappers/FormWrapper'
+export { ContentWrapper } from './objects/wrappers/ContentWrapper'
+export { HeaderWrapper } from './objects/wrappers/HeaderWrapper'
